Fix bug in signExtend32To64
Add 'L' for long to the constant 0x0000_0000_FFFF_FFFF  -> 0x0000_0000_FFFF_FFFFL .
The leading zeros are strip and with out the 'L'  the value is an integer.   ;-(    This causes the bit wise and operator to not clear the  high 32 bits.    This causes a failure.

+++ src/test/groovy/happynewmoonwithreport/type/IntWasmTest.groovy
@@ -104,7 +104,7 @@ class IntWasmTest extends Specification {
 		Long input = new Long(input0);
 
 		when: " extend to signed signed 64"
-		int actual = IntWasm.signExtend32To64(input);
+		long actual = IntWasm.signExtend32To64(input);
 
 
 		then: "Test"
